INSP: add `Trait implementation` inspection
Extract some error annotations about trait implementation into separate inspection.
It should allow users disable and customize this error annotations.
