+++ src/main/kotlin/org/rust/ide/annotator/RsErrorAnnotator.kt
@@ -298,46 +298,10 @@ class RsErrorAnnotator : RsAnnotatorBase(), HighlightRangeExtension {
             !impl.isUnsafe && !trait.isUnsafe && impl.excl == null && attrRequiringUnsafeImpl != null ->
                 RsDiagnostic.TraitMissingUnsafeImplAttributeError(traitRef, attrRequiringUnsafeImpl).addToHolder(holder)
         }
-        val implInfo = TraitImplementationInfo.create(trait, impl) ?: return
-
-        if (implInfo.missingImplementations.isNotEmpty()) {
-            val missing = implInfo.missingImplementations.map { it.name }.namesList
-            RsDiagnostic.TraitItemsMissingImplError(impl.impl, impl.typeReference ?: impl.impl, missing, impl)
-                .addToHolder(holder)
-        }
-
-        for (member in implInfo.nonExistentInTrait) {
-            RsDiagnostic.UnknownMethodInTraitError(member.nameIdentifier!!, member, traitName)
-                .addToHolder(holder)
-        }
-
-        for ((imp, dec) in implInfo.implementationToDeclaration) {
-            if (imp is RsFunction && dec is RsFunction) {
-                checkTraitFnImplParams(holder, imp, dec, traitName)
-            }
-        }
     }
 
     private fun checkTypeAlias(holder: AnnotationHolder, ta: RsTypeAlias) {
         checkDuplicates(holder, ta)
-    }
-
-    private fun checkTraitFnImplParams(holder: AnnotationHolder, fn: RsFunction, superFn: RsFunction, traitName: String) {
-        val params = fn.valueParameterList ?: return
-        val selfArg = fn.selfParameter
-
-        if (selfArg != null && superFn.selfParameter == null) {
-            RsDiagnostic.DeclMissingFromTraitError(selfArg, fn, selfArg).addToHolder(holder)
-        } else if (selfArg == null && superFn.selfParameter != null) {
-            RsDiagnostic.DeclMissingFromImplError(params, fn, superFn.selfParameter).addToHolder(holder)
-        }
-
-        val paramsCount = fn.valueParameters.size
-        val superParamsCount = superFn.valueParameters.size
-        if (paramsCount != superParamsCount) {
-            RsDiagnostic.TraitParamCountMismatchError(params, fn, traitName, paramsCount, superParamsCount)
-                .addToHolder(holder)
-        }
     }
 
     private fun checkUnary(holder: AnnotationHolder, o: RsUnaryExpr) {
@@ -421,9 +385,6 @@ class RsErrorAnnotator : RsAnnotatorBase(), HighlightRangeExtension {
             ?: return false
         return field.parent.parent is RsEnumVariant
     }
-
-    private val Collection<String?>.namesList: String
-        get() = mapNotNull { "`$it`" }.joinToString(", ")
 
     private fun hasResolve(el: RsReferenceElement): Boolean =
         !(el.reference.resolve() != null || el.reference.multiResolve().size > 1)
