+++ src/main/resources/inspectionDescriptions/RsTraitImplementation.html
@@ -1,0 +1,1 @@
+Reports errors in trait implementation like "Missing members", "Unknown member", etc.
