+++ src/test/kotlin/org/rust/ide/inspections/RsTraitImplementationInspectionTest.kt
@@ -1,0 +1,100 @@
+/*
+ * Use of this source code is governed by the MIT license that can be
+ * found in the LICENSE file.
+ */
+
+package org.rust.ide.inspections
+
+class RsTraitImplementationInspectionTest : RsInspectionsTestBase(RsTraitImplementationInspection()) {
+
+    fun `test self in trait not in impl E0186`() = checkErrors("""
+        trait T {
+            fn ok_foo(&self, x: u32);
+            fn ok_bar(&mut self);
+            fn ok_baz(self);
+            fn foo(&self, x: u32);
+            fn bar(&mut self);
+            fn baz(self, o: bool);
+        }
+        struct S;
+        impl T for S {
+            fn ok_foo(&self, x: u32) {}
+            fn ok_bar(&mut self) {}
+            fn ok_baz(self) {}
+            fn foo<error descr="Method `foo` has a `&self` declaration in the trait, but not in the impl [E0186]">(x: u32)</error> {}
+            fn bar<error descr="Method `bar` has a `&mut self` declaration in the trait, but not in the impl [E0186]">()</error> {}
+            fn baz<error descr="Method `baz` has a `self` declaration in the trait, but not in the impl [E0186]">(o: bool)</error> {}
+        }
+    """)
+
+    fun `test self in impl not in trait E0185`() = checkErrors("""
+        trait T {
+            fn ok_foo(&self, x: u32);
+            fn ok_bar(&mut self);
+            fn ok_baz(self);
+            fn foo(x: u32);
+            fn bar();
+            fn baz(o: bool);
+        }
+        struct S;
+        impl T for S {
+            fn ok_foo(&self, x: u32) {}
+            fn ok_bar(&mut self) {}
+            fn ok_baz(self) {}
+            fn foo(<error descr="Method `foo` has a `&self` declaration in the impl, but not in the trait [E0185]">&self</error>, x: u32) {}
+            fn bar(<error descr="Method `bar` has a `&mut self` declaration in the impl, but not in the trait [E0185]">&mut self</error>) {}
+            fn baz(<error descr="Method `baz` has a `self` declaration in the impl, but not in the trait [E0185]">self</error>, o: bool) {}
+        }
+    """)
+
+    fun `test incorrect params number in trait impl E0050`() = checkErrors("""
+        trait T {
+            fn ok_foo();
+            fn ok_bar(a: u32, b: f64);
+            fn foo();
+            fn bar(a: u32);
+            fn baz(a: u32, b: bool, c: f64);
+            fn boo(&self, o: isize);
+        }
+        struct S;
+        impl T for S {
+            fn ok_foo() {}
+            fn ok_bar(a: u32, b: f64) {}
+            fn foo<error descr="Method `foo` has 1 parameter but the declaration in trait `T` has 0 [E0050]">(a: u32)</error> {}
+            fn bar<error descr="Method `bar` has 2 parameters but the declaration in trait `T` has 1 [E0050]">(a: u32, b: bool)</error> {}
+            fn baz<error descr="Method `baz` has 0 parameters but the declaration in trait `T` has 3 [E0050]">()</error> {}
+            fn boo<error descr="Method `boo` has 2 parameters but the declaration in trait `T` has 1 [E0050]">(&self, o: isize, x: f16)</error> {}
+        }
+    """)
+
+    fun `test absent method in trait impl E0046`() = checkErrors("""
+        trait TError {
+            fn bar();
+            fn baz();
+            fn boo();
+        }
+        <error descr="Not all trait items implemented, missing: `bar`, `boo` [E0046]">impl TError for ()</error> {
+            fn baz() {}
+        }
+    """)
+
+    fun `test unknown method in trait impl E0407`() = checkErrors("""
+        trait T {
+            fn foo();
+        }
+        impl T for () {
+            fn foo() {}
+            fn <error descr="Method `quux` is not a member of trait `T` [E0407]">quux</error>() {}
+        }
+    """)
+
+    fun `test different type items have same name E0046`() = checkErrors("""
+        trait A {
+            type C;
+            const C: i32;
+        }
+        <error descr="Not all trait items implemented, missing: `C` [E0046]">impl A for ()</error> {
+            type C = ();
+        }
+    """)
+}
